Rename networkinfo package to status
