+++ app/src/main/java/com/shapeapp/shape/ui/initialbackendconnection/InitialBackendConnectionFragment.kt
@@ -9,7 +9,7 @@ import androidx.fragment.app.Fragment
 import com.shapeapp.shape.R
 import com.shapeapp.shape.backendcalls.RetrofitBackendClient
 import com.shapeapp.shape.data.entities.Card
-import com.shapeapp.shape.data.network.networkinfo.NetworkInformer
+import com.shapeapp.shape.data.network.status.NetworkInformer
 import kotlinx.android.synthetic.main.fragment_initial_backend_connection.*
 import retrofit2.Call
 import retrofit2.Callback
